requirements.txt
tensorflow>=2.13.0
numpy>=1.21.0
pandas>=1.3.0
pandera>=0.22.1
great-expectations>=0.18.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pytest>=7.0.0
pytest-cov>=3.0.0
mlflow
fastapi
uvicorn
watchdog
requests
pydantic
boto3
psycopg2-binary
python-dotenv
sqlalchemy
apache-airflow
minio
csv2


